📦 chore(package.json): add gsap, paper, and simplex-noise as dependencies 🔥 refactor(App.jsx): remove unused code and import MainNavBar component 🆕 feat(animation/cursor.js): add new cursor animation functionality using gsap and simplex-noise library
✨ feat(icons): add new SVG icon for a right arrow menu alternative

A new SVG icon file has been added to the `src/assets/icons` directory. The file is named `gg_menu-right-alt.svg` and contains the markup for a right arrow menu alternative. The icon has a width of 40 and a height of 41. It consists of three paths filled with white color, representing different parts of the icon.

This new icon is intended to be used as an alternative option for a right arrow menu in the application's user interface.

🎨 style: format SVG path in a more readable way
🐛 fix: correct SVG path syntax error

🎨 feat(icons): add new icon for sun in grayscale

The new icon-sun-g.svg file has been added to the src/assets/icons directory. This SVG file represents a grayscale version of the sun icon. It has a width and height of 35 units and a viewBox of 0 0 35 35. The sun icon is filled with a linear gradient that starts with the color #4858BC and ends with the color #6CC8C2.

This new icon will be used to represent the sun in grayscale in the application's user interface.

🎨 style: format SVG path in a more readable way
🐛 fix: correct SVG path syntax error

🎨 feat(icons): add new sun icon SVG file

The new file `icon-sun.svg` has been added to the `src/assets/icons` directory. This SVG file represents a sun icon with a width of 35 and height of 35. It is filled with white color and has a viewBox of 0 0 35 35. This icon can be used for various purposes in the application.

🎨 style: format SVG path in code snippet for better readability

🔧 chore: update SVG path data in file to fix rendering issue

🔧 chore: update SVG path data in file to fix rendering issue

🎨 chore(MainNavBar.jsx): add MainNavBar component to implement a responsive navigation bar
🐛 fix(MainNavBar.jsx): fix SVG path syntax error in commented code
🔥 refactor(MainNavBar.jsx): remove unused imports and variables
🌈 style(MainNavBar.jsx): update SVG fill color to use a linear gradient animation

🎨 style(navbar): remove commented out fill attributes to improve code readability and maintainability
🔧 chore(navbar): refactor navbar component to use motion animations for button icons to enhance user experience

🎨 style(index.css): add cursor styles for interactive elements to improve user experience
✨ feat(main.jsx): add cursor elements to the main component to enable custom cursor functionality
📦 chore(initPageTransitions.js): add new utility file for initializing page transitions
📦 chore(resizer.js): add new utility file for managing resize callbacks
📦 chore(util.js): add new utility file with commonly used utility functions